interview_assistant_8fold-feature-final-voice-agent/interview_assistant_8fold-feature-final-voice-agent/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-dotenv>=1.0.1
fpdf2>=2.7.7

fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-dotenv>=1.0.1
fpdf2>=2.7.7

# LangChain and LLM Packages
langchain>=0.1.16
langchain-core>=0.1.42
langchain-google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-dotenv>=1.0.1
fpdf2>=2.7.7

fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-dotenv>=1.0.1
fpdf2>=2.7.7

# LangChain and LLM Packages
langchain>=0.1.16
langchain-core>=0.1.42
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# Audio Processing
pydub>=0.25.1
pyaudio>=0.2.13

# Database & Auth
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
argon2-cffi>=23.1.0